+++ lib/helpers/class-sync-orders.php
@@ -133,6 +133,8 @@ class ORDER {
 		$order_data = array(
 			'contactCode'            => $order->get_meta( '_billing_vat' ),
 			'contactName'            => $contact_name,
+			'contactFirstName'       => $order->get_billing_first_name(),
+			'contactLastName'        => $order->get_billing_last_name(),
 			'woocommerceCustomer'    => $order->get_user()->data->user_login,
 			'marketplace'            => 'woocommerce',
 			'woocommerceOrderStatus' => $order->get_status(),
@@ -151,6 +153,7 @@ class ORDER {
 			'datestart'              => strtotime( $order->get_date_created() ),
 			'notes'                  => $order->get_customer_note(),
 			'saleschannel'           => null,
+			'currency'               => get_woocommerce_currency(),
 			'language'               => $doclang,
 			'pmtype'                 => null,
 			'items'                  => array(),
@@ -228,7 +231,8 @@ class ORDER {
 		}
 		// Order Items.
 		foreach ( $order->get_items() as $item_id => $item ) {
-			$product = $item->get_product();
+			$product    = $item->get_product();
+			$product_id = ! empty( $product ) ? $product->get_id() : 0;
 
 			if ( ! empty( $product ) && $product->is_type( 'woosb' ) ) {
 				$woosb_ids   = get_post_meta( $item['product_id'], 'woosb_ids', true );
@@ -282,12 +286,14 @@ class ORDER {
 				$item_rate = ! empty( $item_tax ) ? floor( array_shift( $rates ) ) : 0;
 
 				$item_data = array(
-					'name'     => $item->get_name(),
-					'desc'     => get_the_excerpt( $product->get_id() ),
-					'units'    => $item_qty,
-					'subtotal' => (float) $price_line,
-					'tax'      => $item_rate,
-					'sku'      => ! empty( $product ) ? $product->get_sku() : '',
+					'name'      => $item->get_name(),
+					'desc'      => get_the_excerpt( $product_id ),
+					'units'     => $item_qty,
+					'subtotal'  => (float) $price_line,
+					'tax'       => $item_rate,
+					'sku'       => ! empty( $product ) ? $product->get_sku() : '',
+					'image_url' => get_the_post_thumbnail_url( $product_id, 'post-thumbnail' ),
+					'permalink' => get_the_permalink( $product_id ),
 				);
 
 				// Discount.
@@ -313,8 +319,7 @@ class ORDER {
 				// Taxes.
 				$item_tax  = (float) $shipping_item->get_total_tax();
 				$taxes     = $tax->get_rates( $shipping_item->get_tax_class() );
-				$tax_rates = array_shift( $taxes );
-				$item_rate = ! empty( $item_tax ) ? floor( array_shift( $tax_rates ) ) : 0;
+				$item_rate = ! empty( $item_tax ) && is_array( $item_tax ) ? floor( array_shift( $tax_rates ) ) : 0;
 
 				$fields_items[] = array(
 					'name'     => __( 'Shipping:', 'connect-woocommerce' ) . ' ' . $shipping_item->get_name(),